transactions screen template
