+++ app/src/main/java/com/sparshchadha/expensetracker/feature/bottom_navigation/MainBottomNavigationBarFragment.kt
@@ -22,7 +22,7 @@ import com.sparshchadha.expensetracker.common.utils.Utility.noRippleClickable
 import com.sparshchadha.expensetracker.core.navigation.ExpenseTrackerNavGraph.BottomBarScreenRoutes
 import com.sparshchadha.expensetracker.feature.home.ui.fragment.HomeFragment
 import com.sparshchadha.expensetracker.feature.statistics.StatisticsFragment
-import com.sparshchadha.expensetracker.feature.transactions.TransactionsFragment
+import com.sparshchadha.expensetracker.feature.transactions.presentation.ui.fragment.TransactionsFragment
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
