+++ app/src/main/java/com/sparshchadha/expensetracker/feature/home/ui/compose/components/TodaysExpenses.kt
@@ -22,6 +22,7 @@ import com.sparshchadha.expensetracker.common.utils.AppColors
 import com.sparshchadha.expensetracker.common.utils.Dimensions
 import com.sparshchadha.expensetracker.common.utils.FontSizes
 import com.sparshchadha.expensetracker.feature.expense.domain.entity.ExpenseEntity
+import com.sparshchadha.expensetracker.feature.expense.presentation.ui.components.ExpenseCard
 
 @Composable
 fun ExpenseListHandler(
